Rewrite project with better internal structure
To make maintenance of this project easier a rewrite is badly needed.
Previous code lacks clarity and tightly-coupled. In this change a
better introduced:

* Clean stages: Load, Verify, Format.
* Separate types that are not coupled. Certificate, Report, and
  Formatter handle their own parts.
* Simpler output without unnecessary fanciness of lipgloss library.

The output format has changed but I believe it's for good: less emojis,
less colors. Just a straightforward output with a touch of color for
status. Tests and golden files were updated accordingly.